move units.js to svgcanvas package (while making it accessible by static methods)
